fix(raaga): add Math.round() to processing_time_ms; harden Rust deserializer to f64
Raaga engine returned a raw float from performance.now() while all other
TS engines use Math.round(). The Rust TsEngineResponse struct expected u64,
causing a serde deserialization error and BRIDGE_ERROR 500 on every raaga call.

Fix:
- ts-engines/raaga/engine.ts: wrap processing_time_ms in Math.round()
- crates/noesis-bridge/src/ts_client.rs: change processing_time_ms u64 → f64
  so the bridge tolerates fractional milliseconds from any TS engine going forward

Closes #801

+++ crates/noesis-bridge/src/ts_client.rs
@@ -34,8 +34,8 @@ pub struct TsEngineResponse {
     pub witness_prompts: Vec<WitnessPrompt>,
     /// ISO timestamp when calculation completed
     pub calculated_at: String,
-    /// Processing time in milliseconds
-    pub processing_time_ms: u64,
+    /// Processing time in milliseconds (float-safe: TS engines may return fractional ms)
+    pub processing_time_ms: f64,
 }
 
 /// A prompt for self-reflection/witnessing
